reinstall

+++ shell/arch.sh
@@ -73,7 +73,7 @@ open_ssh() {
 continue_install() {
     check_chroot
 
-    if [ "$in_chroot" != 1 ] && [ -f /mnt$user_var_file ]; then
+    if [ "$do_reinstall" != 1 ] && [ "$in_chroot" != 1 ] && [ -f /mnt$user_var_file ]; then
         rsync -t /mnt$user_var_file $user_var_file
     else
         touch $user_var_file
